Switch to a "fork and exec" model of running host commands
This way we can re-use the standard ProcessHandleImpl to monitor
for the host command process finishing instead of getting a
notification back on the main thread. This should fix a thread
deadlock when one monitors the input streams from the main thread.

Another benefit of this way is that it keeps all the glib/dbus
stuff isolated in the forked child process, allowing it to work
in non-SWT applications.

Signed-off-by: Mat Booth <mat.booth@redhat.com>